external/webkit/LayoutTests/platform/android/webgl-conformance/dumpAsText/conformance/textures/texture-complete-expected.txt

 
Checks that a texture that is not -texture-complete- does not draw if filtering needs mips

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS texture that is not -texture-complete- when TEXTURE_MIN_FILTER not NEAREST or LINEAR should draw with 0,0,0,255
PASS successfullyParsed is true

TEST COMPLETE
